requirements.txt
async-timeout==4.0.2
click==8.1.3
Deprecated==1.2.13
Flask==2.2.2
gitdb==4.0.9
GitPython==3.1.29
greenlet==2.0.1
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysqlclient==2.1.1
packaging==21.3
pyparsing==3.0.9
redis==4.3.4
smmap==5.0.0
SQLAlchemy==1.4.43
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.10.0
